thumb arreglado

+++ modules/uploads.py
@@ -2,9 +2,8 @@ from os import unlink
 from requests import get
 from PIL import Image
 from moviepy.editor import VideoFileClip
-
-
-
+from unicodedata import normalize
+import re
 
 def upload_video(app, message_id, data):
     if isinstance(data, list):
@@ -20,19 +19,21 @@ def upload_video(app, message_id, data):
     
 def up(app, message_id, data):
     file = data["file_path"]
+    title = cleanString(data['title'])
+
     if data['description'] == '':
         desc = " " 
     else:
         desc = data['description']
         
-    caption = f"**--Nombre:-- __{data['title']}__\n\n--Descripcion:-- __{desc}__**"
+    caption = f"**--Nombre:-- __{title}__\n\n--Descripcion:-- __{desc}__**"
     
     if len(caption) > 1024:
         caption = tuple(caption)
         caption = "".join(caption[:1024])
     
-    sms = app.send_message(message_id, f"**🚀 Subiendo video: `{data['title']}`**")
-    thumb = download_thumb(url=data['thumb'], name=data['title'])
+    sms = app.send_message(message_id, f"**🚀 Subiendo video: `{title}`**")
+    thumb = download_thumb(url=data['thumb'], name=title)
     video = VideoFileClip(file)
     
     duration = int(video.duration)
@@ -40,19 +41,24 @@ def up(app, message_id, data):
     app.send_video(message_id, file, caption=caption, thumb=thumb, duration=duration)
     sms.delete()
     unlink(file)
-    unlink(thumb)
+
+    if "Designer.png" not in thumb:
+        unlink(thumb)
     
     
 
     
 def upload_audio(app, message_id, data):
-    sms = app.send_message(message_id, f"**🚀 Subiendo audio: `{data['title']}`**")
+    title = cleanString(data['title'])
+    sms = app.send_message(message_id, f"**🚀 Subiendo audio: `{title}`**")
     print("Subiendo audio")
-    thumb = download_thumb(url=data['thumb'], name=data['title'])
+    thumb = download_thumb(url=data['thumb'], name=title)
     app.send_audio(message_id, data['file_path'], thumb=thumb)
     sms.delete()
     unlink(data["file_path"])
-    unlink(thumb)
+    
+    if "Designer.png" not in thumb:
+        unlink(thumb)
     
     
     
@@ -70,7 +76,11 @@ def download_thumb(url, name):
     except Exception as e:
         print(e)
         img_name = "Designer.png"
-    
-        
-    
     return img_name
+
+
+def cleanString(string:str) -> str:
+    text = normalize("NFKD", string).encode("ascii", "ignore").decode("utf-8", "ignore")
+    text = re.sub(r'[\[\]\(\)]', '', text)
+    text = text.replace("/", " ")
+    return text
